requirements.txt
langchain_core
python-dotenv
langchain_openai
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette